:bug: Fix - les markers ne s'affichaient plus + ApiDAO delete
